pyproject.toml
[tool.poetry]
name = "fabliteelectronics"
version = "0.1.0"
description = ""
authors = ["DarthZiel <Studerzhan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
drf-yasg = "^1.21.7"
gunicorn = "^22.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
